requirements.txt
jinja2==3.1.2
pyyaml==6.0
sqlalchemy==2.0.13
loguru==0.7.0
psycopg2-binary==2.9.3